ServerReady

+++ Domain/OrderBatchs/BatchWeightResults/BatchWeighingResult.cs
@@ -12,11 +12,11 @@ namespace Domain.OrderBatchs.BatchWeightResults
         public Guid BatchWeighingResultId { get; private set; }
         public Guid OrderBatchId { get; private set; }
         public string MaterialId { get; private set; }
-        public decimal TargetKg { get; private set; }
-        public decimal ActualKg { get; private set; }
-        public decimal DeviationKg { get; private set; }
-        public decimal ToleranceMinKg { get; private set; }
-        public decimal ToleranceMaxKg { get; private set; }
+        public float TargetKg { get; private set; }
+        public float ActualKg { get; private set; }
+        public float DeviationKg { get; private set; }
+        public float ToleranceMinKg { get; private set; }
+        public float ToleranceMaxKg { get; private set; }
         public bool IsWithinTolerance { get; private set; }
         public DateTime CapturedAt { get; private set; } = DateTime.Now;
         public OrderBatch OrderBatch { get; private set; }
@@ -33,10 +33,10 @@ namespace Domain.OrderBatchs.BatchWeightResults
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
             Guid orderBatchId,
             string materialId,
-            decimal targetKg,
-            decimal actualKg,
-            decimal toleranceMinKg,
-            decimal toleranceMaxKg)
+            float targetKg,
+            float actualKg,
+            float toleranceMinKg,
+            float toleranceMaxKg)
         {
             OrderBatchId = orderBatchId;
             MaterialId = materialId;
